refactor: dto는 사용하는 측에서 생성하게 변경
- 사용하는 측은 controller 혹은 상위 계층의 모델임
- 상위 계층의 모델은 player 와 players 중 players를 의미함
