+++ src/main/java/blackjack/dto/DealerFinalCardsOutcome.java
@@ -1,6 +1,7 @@
-package blackjack.model.dealer;
+package blackjack.dto;
 
 import blackjack.model.card.Card;
+import blackjack.model.dealer.Dealer;
 import java.util.List;
 
 public record DealerFinalCardsOutcome(List<Card> cards, int totalScore) {
